Move tests requiring PHP ≥ 8.1 to their own directory
