+++ composer.json
@@ -66,7 +66,8 @@
         "psr-4": {
             "Doctrine\\Tests\\": "tests/Doctrine/Tests",
             "Doctrine\\StaticAnalysis\\": "tests/Doctrine/StaticAnalysis",
-            "Doctrine\\Performance\\": "tests/Doctrine/Performance"
+            "Doctrine\\Performance\\": "tests/Doctrine/Performance",
+            "Doctrine\\Tests_PHP81\\": "tests_php81"
         }
     },
     "bin": ["bin/doctrine"],
